refactor: use fluent builders from lib

+++ src/test/java/com/sipgate/li/simulator/e2e/E2ETest.java
@@ -545,7 +545,7 @@ public class E2ETest {
   private static void assertErrorResponse(
     final ErrorResponse response,
     final RequestMessageType type,
-    final BigInteger code,
+    final long code,
     final String description
   ) {
     assertThat(response.getRequestMessageType()).isEqualTo(type);
